requirements.txt
streamlit
openai
python-pptx
requests